backend/ninepm/app/ninepm_mobile_api/base/serializers/abstract_serializer.py
__all__ = [
    "AbstractSerializer",
]

from rest_framework import serializers


class AbstractSerializer(serializers.Serializer):

    def update(self, instance, validated_data):
        pass

    def create(self, validated_data):
        pass
