backend/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crc16==0.1.1
cryptography==41.0.0
dataclasses==0.6
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-redis==5.2.0
djangorestframework==3.14.0
redis==4.5.5
drf-yasg==1.21.5
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==18.9.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.1
paramiko==3.2.0
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.4.2
redis==4.5.5
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
uWSGI==2.0.21
zipp==3.15.0
firebase-admin==6.1.0
pika==1.3.2
